estilos para redirect, funcion correo Globlal
